set up mock data

+++ cardgame/src/main/java/com/yrgo/sp/cardgame/data/CardRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.yrgo.sp.cardgame.domain.Card;
 
-public interface CardRepository extends JpaRepository<Card, Long>{
+public interface CardRepository extends JpaRepository<Card, Integer>{
 	public List<Card> findAll();
 }
